Add test case for additional root apps config map
The list of teams in the config map should never contain the team owning
the cluster, since that team keeps the "default" root app named `root`
and the default app project named `syn`.

Additionally, we don't want to generate a root app and app project for
teams which are listed in the parameter but which don't have any
component instances assigned.

The test case is setup to have a structure in parameter `syn` which
covers the cases outlined above.

Rendered from template version: main (ff9d5a8)